# some more wire formatting improvements