… projects/RabbitMQ.Client/client/impl/WireFormatting.cs