## Proposed Changes

Benchmarks:
- Added size benchmarks
- Added DataTypeFieldSerialization for table / array serialization

Optimizations
- Aligned some optimization from table to array
- Rearranged (or factored out into slow method) switches for types for read/write/size 
- Use span element ref to reduce code size 
- Use Int instead of byte for bool conversions

Breaking changes
- ReadArray now returns null instead of empty List => Can be excluded if not wanted for backport

## Types of Changes

- [ ] Bug fix (non-breaking change which fixes issue #NNNN)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause an observable behavior change in existing systems)
- [ ] Documentation improvements (corrections, new content, etc)
- [ ] Cosmetic change (whitespace, formatting, etc)

## Checklist

- [x] I have read the `CONTRIBUTING.md` document
- [x] I have signed the CA (see https://cla.pivotal.io/sign/rabbitmq)
- [ ] All tests pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
- [ ] Any dependent changes have been merged and published in related repositories